+++ src/main/java/life/senlin/communication/Controller/QuestionController.java
@@ -19,7 +19,7 @@ public class QuestionController {
     QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable("id") Integer id,
+    public String question(@PathVariable("id") Long id,
                            Model model){
         QuestionDTO questionDTO = questionService.findById(id);
         //累加阅读数
